Fill some repositories and create tests for them

+++ src/main/java/com/apu/olxcrawler/repository/PhoneRepository.java
@@ -13,10 +13,10 @@ import com.apu.olxcrawler.repository.entity.PhoneEntity;
  */
 public interface PhoneRepository {
     
-    PhoneEntity getPhoneById(Long id);
+    PhoneEntity getPhoneById(Integer id);
     
     Integer getIdByPhoneNumber(String phoneNumber);
     
-    Long add(String phoneNumber);
+    Integer add(String phoneNumber);
     
 }
